backend/requirements.txt
fastapi==0.110.0
uvicorn[standard]==0.24.0
sqlalchemy==2.0.36
alembic==1.12.1
pydantic==2.9.2
pydantic-settings==2.5.2
python-multipart==0.0.6
passlib[bcrypt]==1.7.4
python-jose[cryptography]==3.3.0
pytest==7.4.3
pytest-asyncio==0.21.1
httpx==0.25.2
aiosqlite==0.19.0
python-magic==0.4.27
greenlet==3.0.3
openpyxl==3.1.5
